Change primitives to their object counter-parts, Transfer CRUD WIP

+++ src/main/java/ru/cft/igoshin/api/dto/wallet/WalletResponse.java
@@ -4,8 +4,8 @@ import jakarta.validation.constraints.NotNull;
 
 public record WalletResponse(
         @NotNull
-        int number,
+        Long number,
         @NotNull
-        int balance
+        Long balance
 ) {
 }
