+++ src/main/java/ru/cft/igoshin/core/model/Wallet.java
@@ -14,10 +14,10 @@ public class Wallet {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, updatable = false)
-    private int id;
+    private Long id;
 
     @Column(name = "balance", nullable = false)
-    private int balance;
+    private Long balance;
 
     @OneToOne(mappedBy = "wallet", optional = false)
     private User user;
